REF: implement a new binary-based data rate file format
Reworks all the rate related code so that it uses the binary format, deprecates
one-to-one worker channel since it is not needed on this new format.

Refactor the aggregator functionality so it does not rely on the CSV files to
aggregate the data and rebuilds the latency information by joining the HDR
files already present on the report directory.

+++ maestro-plotters/maestro-plotter-rate/src/main/java/org/maestro/plotter/rate/RateData.java
@@ -58,9 +58,9 @@ public class RateData implements ReportData {
         return list;
     }
 
-    public List<Integer> getRateValues()
+    public List<Long> getRateValues()
     {
-        return recordSet.stream().mapToInt(RateRecord::getCount).boxed().collect(Collectors.toList());
+        return recordSet.stream().mapToLong(RateRecord::getCount).boxed().collect(Collectors.toList());
     }
 
     @Override
